style: cleanup pr

+++ pkg/twoqueue/2q.go
@@ -307,7 +307,7 @@ func (c *TwoQueueCache[K, V]) Len() int {
 // For generic caches, this returns 0 as the size cannot be determined without type information.
 // Specialized implementations should override this method.
 func (c *TwoQueueCache[K, V]) SizeBytes() int64 {
-	return int64(c.frequent.SizeBytes() + int64(size.Of(c.recent.cache)) + int64(size.Of(c.ghost.cache)))
+	return c.frequent.SizeBytes() + int64(size.Of(c.recent.cache)) + int64(size.Of(c.ghost.cache))
 }
 
 // ensureRecentSpace makes room in the recent cache by evicting items when necessary.
